Add LIB to Cryptarchia
Add a tracker for the Latest Immutable Block (LIB) in Cryptarchia
as described in the spec + related tests.
Does not prune forks as soon as a block is applied (like the spec)
as that needs to be done in tandem with storage and it's still a
WIP at the moment. A future update might bring the implementation
closer to the reference spec.