Partial update to PoLv2
This commit paves the ground for PoLv2 by changing the public and
private inputs to the proof and related traits. It's not a
complete implementation and does not attempt to build a valid proof
as that would require the new mantle note format. Instead, the
purpose of this PR is to split the work needed for the new mantle
changes and enable the note changes.